requirements.txt
asgiref==3.11.0
Django==5.2
djangorestframework==3.16.1
pillow==12.1.0
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.3
